debuged and corrected results of calculating

+++ include/ComputingOfBeauty.hpp
@@ -5,17 +5,17 @@
 
 class ComputingOfBeauty{
     private:
-        int f_base{2};
-        int f_num_len{1};
+        unsigned int f_base{2};
+        unsigned int f_num_len{1};
 
     public:
-        ComputingOfBeauty(int base, int num_len);
+        ComputingOfBeauty(unsigned int base, unsigned int num_len);
         ~ComputingOfBeauty();
 
-        int sum_of_digits(unsigned int val);
-        unsigned long int count_num_of_beauties();
-        int get_base();
-        int get_num_len();
+        unsigned int sum_of_digits(unsigned int val);
+        unsigned long long count_num_of_beauties();
+        unsigned int get_base();
+        unsigned int get_num_len();
 };
